### 2019-03-06 #1300 @arfon
:wave: @shawnbanasick - this PR will fix your paper compilation issues https://github.com/shawnbanasick/kade/pull/1

/ cc @openjournals/joss-eics
