### 2019-03-14 #1300 @labarba
👋  Hi @alexhanna! Welcome to the JOSS Editorial Board — we think this submission is aligned with your interest and expertise. Can you handle as editor?
