### 2019-03-14 #1300 @alexhanna
Yes, I can take this on.

On Thu, Mar 14, 2019 at 11:36 AM Lorena A. Barba <notifications@github.com>
wrote:

> 👋 Hi @alexhanna <https://github.com/alexhanna>! Welcome to the JOSS
> Editorial Board — we think this submission is aligned with your interest
> and expertise. Can you handle as editor?
>
