### 2019-04-01 #1300 @davekinkead
@alexhanna I can review this later this week
