### 2019-04-01 #1360 @whedon
[REVIEW]: KADE: A desktop application for Q methodology
**Submitting author:** @shawnbanasick (<a href="http://orcid.org/0000-0003-2696-5874">Shawn Banasick</a>)
**Repository:** <a href="https://github.com/shawnbanasick/kade" target ="_blank">https://github.com/shawnbanasick/kade</a>
**Version:** v1.0.0
