### 2019-04-05 #1360 @davekinkead
@alexhanna @shawnbanasick outstanding review issues raised here https://github.com/shawnbanasick/kade/issues

I've only been able to review the software on MacOS (demo works as per readme) so can't comment to its functionality on Windows or Linux.
