### 2019-04-21 #1360 @alexhanna
You'll need to get a DOI for your repository, @shawnbanasick. You'll need to create an archive (on Zenodo, figshare, or somewhere else, if you haven't already) and post the archive DOI here.
