### 2019-04-23 #1360 @alexhanna
Hi @openjournals/joss-eics, this is ready for acceptance.
