### 2019-04-23 #1360 @danielskatz
Thanks for reviewing @davekinkead and for editing @alexhanna
